requirements-ci.txt
jsonschema>=4.22.0,<5
PyYAML>=6.0.1,<7